HashMap for Costs; Detailed Requirements Added
Added a HashMap that costs from topo.txt will be recorded in. The key is in the form of (register 1, register 2),  with the cost between the two as the value.
Requirements were added that explained the overall goal of the program, as well as some future thoughts for how I want to approach creating the forwarding table.