Update to Algorithm
Changed algorithm entirely. It is much more simple now, and also produces an output of the different required numbers. Still need to tweak some things with the algorithm; mainly, need to create a method that updates past nodes to find more efficient paths based on new knowledge. After that, the forwarding table is all that is left.